+++ src/main/java/org/example/rm_back/admin/service/AdminService.java
@@ -5,7 +5,7 @@ import org.example.rm_back.admin.dto.AdminRequestDto;
 import org.example.rm_back.admin.dto.AdminResponseDto;
 import org.example.rm_back.admin.dto.AdminResponseDto.Add;
 import org.example.rm_back.admin.entity.Admin;
-import org.example.rm_back.admin.entity.AdminRole;
+import org.example.rm_back.admin.enums.AdminRole;
 import org.example.rm_back.admin.repository.AdminRepository;
 import org.example.rm_back.common.BCryptEncryptor;
 import org.example.rm_back.global.exception.ApiException;
